Apply the functional options design patterns
For each option function provided to the NewHandler,
we call that function by passing in a pointer to the handler struct
that was just declared.

Using this pattern, we can make API that has is highly configurable,
can grow over time, self documenting, safe for newcorners and never
requires a nil or an empty value to keep the compiler happy.

Signed-off-by: Ronald Ngounou <74538524+ronaldngounou@users.noreply.github.com>